large behavior change here.  i found that if the IMAP server is *dropping* packets on port 143 (e.g., like espresso.foo-projects.org does), it takes over three minutes for the connection to time out.  this is unacceptable: if the panel tries to quit during this time, it'll hang until the connection times out (while in g_thread_join()).  so now we set the socket non-blocking before connect(), and then select() on the socket with a low timeout, over and over, until either the connection succeeds or fails.  we also check the GAsyncQueue between each select() to see if we've been instructed to quit.
also, i implemented a sort of "caching": if port 143 fails more than 3 times
in a row, it'll temporarily (for 5 hours) stop checking port 143, and just
go for 993 with a TLS negotiation.  if both 143 and 993 fail at the same time,
we assume it's a transient error and try the normal sequence next time.

whew.


(Old svn revision: 1551)