use STATUS instead of EXAMINE+SEARCH for IMAP.
should be faster and (hopefully) less error-prone, and avoids a lot of
allocations since the response length is more or less fixed (and small).
also better handle errors.  if we get "#### NO" from the STATUS request,
just return 0 new messages.  that can happen if, for example, a folder
is selected in mailwatch that is later deleted or unsubscribed.


(Old svn revision: 3915)